sentiment-analysis/requirements.txt
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
matplotlib==3.9.2
python-dotenv==1.0.1
torch==2.5.1
transformers==4.46.2
datasets==3.1.0
accelerate==1.1.1
evaluate==0.4.3
tqdm==4.67.0